+++ discordia/classes/color.lua
@@ -72,7 +72,7 @@ function Color.__div(a, b)
 end
 
 function Color:toHex()
-	return string.format('%02X%02X%02X', self.r, self.g, self.b)
+	return string.format('0x%02X%02X%02X', self.r, self.g, self.b)
 end
 
 function Color:toDec()