+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 *Backwards compatibility not guaranteed until after version 1.0.0*
 
-### Future
+### 0.5.3
 - Added "0x" to Color:toHex() output
 - Added Permissions:toHex() method
 - Fixed issue where server owner was nil